Issue #533: General Purpose Meter & Measurement

+++ app/src/main/java/org/supla/android/lib/SuplaChannelBase.java
@@ -27,7 +27,7 @@ public class SuplaChannelBase {
   public int Func;
   public int AltIcon;
   public int UserIcon;
-  public int Flags;
+  public long Flags;
   public boolean OnLine;
   public String Caption;
